removed deprecated stuff
