--- actsims/binningTest
@@ -1,80 +1,0 @@
-0.0  140.0  70.0
-141.0  240.0  190.5
-241.0  340.0  290.5
-341.0  440.0  390.5
-441.0  540.0  490.5
-541.0  640.0  590.5
-641.0  740.0  690.5
-741.0  840.0  790.5
-841.0  940.0  890.5
-941.0  1040.0  990.5
-1041.0  1140.0  1090.5
-1141.0  1240.0  1190.5
-1241.0  1340.0  1290.5
-1341.0  1440.0  1390.5
-1441.0  1540.0  1490.5
-1541.0  1640.0  1590.5
-1641.0  1740.0  1690.5
-1741.0  1840.0  1790.5
-1841.0  1940.0  1890.5
-1941.0  2040.0  1990.5
-2041.0  2140.0  2090.5
-2141.0  2340.0  2240.5
-2341.0  2540.0  2440.5
-2541.0  2740.0  2640.5
-2741.0  2940.0  2840.5
-2941.0  3340.0  3140.5
-3341.0  3740.0  3540.5
-3741.0  4140.0  3940.5
-4141.0  4540.0  4340.5
-4541.0  4940.0  4740.5
-4941.0  5840.0  5390.5
-5841.0  6740.0  6290.5
-6741.0  7640.0  7190.5
-7641.0  8540.0  8090.5
-8541.0  9440.0  8990.5
-9441.0  10340.0  9890.5
-10341.0  11240.0  10790.5
-11241.0  12140.0  11690.5
-12141.0  13040.0  12590.5
-13041.0  13940.0  13490.5
-13941.0  14840.0  14390.5
-14841.0  15740.0  15290.5
-15741.0  16640.0  16190.5
-16641.0  17540.0  17090.5
-17541.0  18440.0  17990.5
-18441.0  19340.0  18890.5
-19341.0  20240.0  19790.5
-20241.0  21140.0  20690.5
-21141.0  22040.0  21590.5
-22041.0  22940.0  22490.5
-22941.0  23840.0  23390.5
-23841.0  24740.0  24290.5
-24741.0  25640.0  25190.5
-25641.0  26540.0  26090.5
-26541.0  27440.0  26990.5
-27441.0  28340.0  27890.5
-28341.0  29240.0  28790.5
-29241.0  30140.0  29690.5
-30141.0  31040.0  30590.5
-31041.0  31940.0  31490.5
-31941.0  32840.0  32390.5
-32841.0  33740.0  33290.5
-33741.0  34640.0  34190.5
-34641.0  35540.0  35090.5
-35541.0  36440.0  35990.5
-36441.0  37340.0  36890.5
-37341.0  38240.0  37790.5
-38241.0  39140.0  38690.5
-39141.0  40040.0  39590.5
-40041.0  40940.0  40490.5
-40941.0  41840.0  41390.5
-41841.0  42740.0  42290.5
-42741.0  43640.0  43190.5
-43641.0  44540.0  44090.5
-44541.0  45440.0  44990.5
-45441.0  46340.0  45890.5
-46341.0  47240.0  46790.5
-47241.0  48140.0  47690.5
-48141.0  49040.0  48590.5
-49041.0  49940.0  49490.5
